flatpak: Set Junction as default http(s) handler in flatpak
When opening a uri, GTK tries locally first in the sandbox and falls back to the portal.

Since we expose host applications to the sandbox, GTK will take the first handler for http(s) and launch it.

The application will fail to run from the sandbox which breaks Gtk.show_uri, GtkLinkButton and so on.

To workaround the issue set Junction as the default http(s) handler within the sandbox.

See https://github.com/flathub/flathub/pull/2585#issuecomment-945092936
See https://gitlab.gnome.org/GNOME/glib/-/issues/1960