flatpak: Fix application listing by patching GLib
There are too many issues with Exec and TryExec. For example; on Arch Firefox has /usr/lib/firefox/firefox, Chromium /usr/bin/chromium, ...

Right now it only works if the name of program is present in $PATH so it's also limited for custom scripts and so on.

This also speeds things up a bit and simplify code.