Open the context menu on Alt-Return
Also on Alt-Space but that's usually handled by Mutter.
Also on Alt-Number, and additionally handle Ctrl-Number.

Closes #126

Currently the number keys are handled by the window:
pressing one will focus the corresponding tile
and releasing the key will activate that tile.
Return and Space are handled by the tile itself.

This change moves responsibility for handling
key-released events of number keys
from the window to the tiles.
The key-pressed event focuses the tile
so it can handle the subsequent key-released event.
The tiles don't know their corresponding number
but since the tile is focused it can assume
that the number key was meant for it.